requirements.txt
amqp==5.0.5
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cheroot==8.4.5
CherryPy==18.6.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dominate==2.6.0
drf-yasg==1.17.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-oauth2-tool==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
Jinja2==2.11.2
kombu==5.0.2
lxml==4.6.1
MarkupSafe==1.1.1
more-itertools==8.6.0
oauth2client==4.1.3
packaging==20.4
pipenv==2018.11.26
portend==2.6
prompt-toolkit==3.0.14
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
pytz==2020.4
redis==3.5.3
requests==2.24.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tempora==4.0.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.0.1
zc.lockfile==2.0
zipp==3.4.0
